requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
jwt==1.3.1
multidict==6.0.5
pycparser==2.22
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
